test(core): split browserslist runtime tests
